+++ packages/simple_graphql_client_gen/test/use_test.dart
@@ -13,6 +13,13 @@ void main() {
     );
   });
 
+  test('empty', () async {
+    expect(
+      await Api.empty(uri, null),
+      const QueryEmpty(),
+    );
+  });
+
   test('account', () async {
     expect(
       (await Api.account(id: const ID('sampleId'), uri, null)).account,